Last Push of the night
The extra strafeLeft did fix the going-right-problem.
We still like to hit walls but we react much fast now (a value change of 25-50 would give us all the time we need). I think I found what was causing the strange running in circles behavior. (other than panic) I was calling a 45-degree left turn inside of a for loop, so I was really calling a 3000+degree turn, hence the circles